Add flash-message toasts, delete guards, and a whole-month invoice shortcut
- Flash toasts (success/error) shared via HandleInertiaRequests and shown
  in the sidebar layout; controllers now guard blocked deletes with an
  error flash.
- Invoices: add a "عرض الشهر كامل" shortcut that fills the date range with
  a picked month's inclusive bounds and views it for the selected dentist.
- Orders create/edit and price-list editor tweaks; dentist request rules.

+++ app/Http/Requests/StoreDentistRequest.php
@@ -26,7 +26,9 @@ class StoreDentistRequest extends FormRequest
             'phone' => ['nullable', 'string', 'unique:dentists,phone'],
             'address' => ['nullable', 'string'],
             'price_list' => ['nullable', 'array'],
-            'price_list.*' => ['numeric', 'min:0'],
+            // Integer to match order items' price rule — a decimal here would
+            // auto-fill into the order form and then fail its validation.
+            'price_list.*' => ['integer', 'min:0'],
         ];
     }
 }
