+++ app/Http/Requests/UpdateDentistRequest.php
@@ -26,7 +26,9 @@ class UpdateDentistRequest extends FormRequest
             'phone' => ['nullable', 'string', \Illuminate\Validation\Rule::unique('dentists')->ignore($this->dentist)],
             'address' => ['nullable', 'string'],
             'price_list' => ['nullable', 'array'],
-            'price_list.*' => ['numeric', 'min:0'],
+            // Integer to match order items' price rule — a decimal here would
+            // auto-fill into the order form and then fail its validation.
+            'price_list.*' => ['integer', 'min:0'],
         ];
     }
 }
